complete integration tests
