--- tests/integration/test_continuous.py
@@ -1,10 +1,0 @@
-from .common import PlaywrightTestCase
-
-
-class TestContinuousIntegration(PlaywrightTestCase):
-    def test_continuous(self):
-        page = self.page
-
-        page.goto(self.url("/"))
-
-        # TODO - keep adding other expect checks
